+++ src/app/academic/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { NoticeLayout } from '@/components/layout/notice-layout';
+import { NoticeLayout } from '@/components/notice/notice-layout';
 import { academicNotices } from '@/data/notices';
 import { Notice } from '@/types/notice';
