+++ src/components/notice/animated-notice-list.tsx
