+++ src/components/notice/notice-header.tsx
